helper functions created
