+++ Plans-shop/Projects/plantsshop.website/src/pages/HomePage/ArrivalsSection/Arrivals.jsx
@@ -10,6 +10,7 @@ import axios from 'axios';
 import "./Arrivals.css";
 import { useNavigate } from 'react-router-dom';
 import { AuthContext } from '../../../context/AuthContext';
+import { handleCartToggle, handleFavoriteToggle, fetchCartItems, handleCloseSnackbar, fetchWishlistItems } from '../../../utils/HelperFunctions';
 
 export default function Arrivals() {
     const [slideIndex, setSlideIndex] = useState(0);
@@ -33,57 +34,9 @@ export default function Arrivals() {
             }
         };
 
-        const fetchCartItems = async () => {
-            try {
-                if (user) {
-                    const response = await axios.get("/src/assets/Constants.json");
-                    const apiBaseUrl = response.data.API_HOST;
-
-                    // Fetch or create cart by customer ID
-                    const cartResponse = await axios.get(`${apiBaseUrl}/api/Cart/getOrCreateCartByCustomerId/${user.id}`);
-                    const currentCartId = cartResponse.data.id;
-
-                    // Fetch cart items by cart ID
-                    const cartItemsResponse = await axios.get(`${apiBaseUrl}/api/CartItem/getByCartId/${currentCartId}`);
-                    setCart(cartItemsResponse.data);
-
-                    // Set cart ID for future use
-                    if (setCartId) {
-                        setCartId(currentCartId);
-                    }
-                }
-            } catch (error) {
-                console.error('Error fetching cart items:', error);
-            }
-        };
-
-        const fetchWishlistItems = async () => {
-            try {
-                if (user) {
-                    const response = await axios.get("/src/assets/Constants.json");
-                    const apiBaseUrl = response.data.API_HOST;
-
-                    // Fetch or create wishlist by customer ID
-                    const wishlistResponse = await axios.get(`${apiBaseUrl}/api/Wishlist/getOrCreateWishlistByCustomerId/${user.id}`);
-                    const currentWishlistId = wishlistResponse.data.id;
-
-                    // Fetch wishlist items by wishlist ID
-                    const wishlistItemsResponse = await axios.get(`${apiBaseUrl}/api/WishlistItems/getByWishlistId/${currentWishlistId}`);
-                    setWishlist(wishlistItemsResponse.data);
-
-                    // Set wishlist ID for future use
-                    if (setWishlistId) {
-                        setWishlistId(currentWishlistId);
-                    }
-                }
-            } catch (error) {
-                console.error('Error fetching wishlist items:', error);
-            }
-        };
-
         fetchProducts();
-        fetchCartItems();
-        fetchWishlistItems();
+        fetchCartItems(user, setCart, setCartId);
+        fetchWishlistItems(user,setWishlist,setWishlistId);
     }, [user, setCartId, setWishlistId, setCart, setWishlist]);
 
     const handleNext = () => {
@@ -107,209 +60,34 @@ export default function Arrivals() {
         });
     }, [slideIndex]);
 
-
-    const handleCloseSnackbar = () => {
-        setSnackbar({ open: false, message: "", severity: "success" });
+    const handleAddToCartToggle = (product) => {
+        handleCartToggle(
+            product,
+            user,
+            cart,
+            cartId,
+            setCart,
+            setCartId,
+            (snackbarMessage) => setSnackbar((prevState) => ({ ...prevState, open: true, message: snackbarMessage })),
+            (snackbarSeverity) => setSnackbar((prevState) => ({ ...prevState, open: true, severity: snackbarSeverity }))
+        );
     };
 
-    const handleCartToggle = async (product) => {
-        if (!user) {
-            alert("Please log in to add items to the cart.");
-            return;
-        }
-
-        try {
-            const response = await axios.get("/src/assets/Constants.json");
-            const apiBaseUrl = response.data.API_HOST;
-
-            // Check if the product is already in the cart
-            const existingCartItem = cart.find((item) => item.product_id === product.id);
-
-            if (existingCartItem) {
-                // If the product is in the cart, remove it
-                await axios.delete(`${apiBaseUrl}/api/CartItem/delete/${existingCartItem.id}`);
-                setCart((prevCart) => prevCart.filter((item) => item.id !== existingCartItem.id));
-
-                // Show snackbar for item removed
-                setSnackbar({ open: true, message: "Item removed from cart", severity: "warning" });
-            } else {
-                // If the product is not in the cart, add it
-                let currentCartId = cartId;
-                if (!currentCartId) {
-                    const cartResponse = await axios.get(`${apiBaseUrl}/api/Cart/getOrCreateCartByCustomerId/${user.id}`);
-                    if (cartResponse.data && cartResponse.data.id) {
-                        currentCartId = cartResponse.data.id;
-                    } else {
-                        const newCartResponse = await axios.post(`${apiBaseUrl}/api/Cart/add`, {
-                            Customer_id: user.id,
-                            Customer: {
-                                first_Name: user.first_Name,
-                                last_Name: user.last_Name,
-                                phone_Number: user.phone_Number,
-                                email: user.email,
-                                password: user.password // Include the password here
-                            }
-                        });
-                        currentCartId = newCartResponse.data.id;
-                    }
-                    setCartId(currentCartId);
-                }
-
-                // Fetch subcategory and category details
-                const subcategoryResponse = await axios.get(`${apiBaseUrl}/api/SubCategories/${product.sub_categories_id}`);
-                const subCategory = subcategoryResponse.data;
-                const categoryResponse = await axios.get(`${apiBaseUrl}/api/Category/${subCategory.category_id}`);
-                const category = categoryResponse.data;
-
-                // Calculate total
-                const quantity = 1;  // Default quantity
-                const total = product.price * quantity;
-
-                // Add item to cart with all necessary details
-                const cartItem = {
-                    cart_id: currentCartId,
-                    product_id: product.id,
-                    quantity: quantity,
-                    total: total,
-                    Product: {
-                        id: product.id,
-                        name: product.name,
-                        price: product.price,
-                        image_url: product.image_url,
-                        sub_categories_id: product.sub_categories_id,
-                        SubCategories: {
-                            id: subCategory.id,
-                            name: subCategory.name,
-                            Category: {
-                                id: category.id,
-                                name: category.name
-                            }
-                        }
-                    },
-                    Cart: {
-                        id: currentCartId,
-                        Customer: {
-                            id: user.id,
-                            first_Name: user.first_Name,
-                            last_Name: user.last_Name,
-                            phone_Number: user.phone_Number,
-                            email: user.email,
-                            password: user.password // Include the password here
-                        }
-                    }
-                };
-
-                await axios.post(`${apiBaseUrl}/api/CartItem/add`, cartItem);
-                setCart((prevCart) => [...prevCart, cartItem]);
-
-                // Show snackbar for item added
-                setSnackbar({ open: true, message: "Item added to cart", severity: "success" });
-            }
-        } catch (error) {
-            console.error("Error adding/removing item to/from cart:", error);
-        }
+    const handleAddToWishlistToggle = (product) => {
+        handleFavoriteToggle(
+            product,
+            user,
+            wishlist,
+            wishlistId,
+            setWishlist,
+            setWishlistId,
+            (snackbarState) => setSnackbar((prevState) => ({ ...prevState, ...snackbarState }))
+        );
     };
-
-
-    const handleFavoriteToggle = async (product) => {
-        if (!user) {
-            alert("Please log in to add items to the wishlist.");
-            return;
-        }
-
-        try {
-            const response = await axios.get("/src/assets/Constants.json");
-            const apiBaseUrl = response.data.API_HOST;
-
-            // Check if the product is already in the wishlist
-            const existingWishlistItem = wishlist.find((item) => item.product_id === product.id);
-
-            if (existingWishlistItem) {
-                // If the product is in the wishlist, remove it
-                await axios.delete(`${apiBaseUrl}/api/WishlistItems/delete/${existingWishlistItem.id}`);
-                setWishlist((prevWishlist) => prevWishlist.filter((item) => item.id !== existingWishlistItem.id));
-
-                // Show snackbar for item removed
-                setSnackbar({ open: true, message: "Item removed from wishlist", severity: "warning" });
-            } else {
-                // If the product is not in the wishlist, add it
-                let currentWishlistId = wishlistId;
-                if (!currentWishlistId) {
-                    const wishlistResponse = await axios.get(`${apiBaseUrl}/api/Wishlist/getOrCreateWishlistByCustomerId/${user.id}`);
-                    if (wishlistResponse.data && wishlistResponse.data.id) {
-                        currentWishlistId = wishlistResponse.data.id;
-                    } else {
-                        const newWishlistResponse = await axios.post(`${apiBaseUrl}/api/Wishlist/add`, {
-                            Customer_id: user.id,
-                            Customer: {
-                                first_Name: user.first_Name,
-                                last_Name: user.last_Name,
-                                phone_Number: user.phone_Number,
-                                email: user.email,
-                                password: user.password
-                            }
-                        });
-                        currentWishlistId = newWishlistResponse.data.id;
-                    }
-                    if (setWishlistId) {
-                        setWishlistId(currentWishlistId);
-                    }
-                }
-
-                // Fetch subcategory and category details
-                const subcategoryResponse = await axios.get(`${apiBaseUrl}/api/SubCategories/${product.sub_categories_id}`);
-                const subCategory = subcategoryResponse.data;
-                const categoryResponse = await axios.get(`${apiBaseUrl}/api/Category/${subCategory.category_id}`);
-                const category = categoryResponse.data;
-
-                // Add item to wishlist with all necessary details
-                const wishlistItem = {
-                    wishlist_id: currentWishlistId,
-                    product_id: product.id,
-                    Product: {
-                        id: product.id,
-                        name: product.name,
-                        price: product.price,
-                        image_url: product.image_url,
-                        sub_categories_id: product.sub_categories_id,
-                        SubCategories: {
-                            id: subCategory.id,
-                            name: subCategory.name,
-                            Category: {
-                                id: category.id,
-                                name: category.name
-                            }
-                        }
-                    },
-                    Wishlist: {
-                        id: currentWishlistId,
-                        Customer: {
-                            id: user.id,
-                            first_Name: user.first_Name,
-                            last_Name: user.last_Name,
-                            phone_Number: user.phone_Number,
-                            email: user.email,
-                            password: user.password
-                        }
-                    }
-                };
-
-                await axios.post(`${apiBaseUrl}/api/WishlistItems/add`, wishlistItem);
-                setWishlist((prevWishlist) => [...prevWishlist, wishlistItem]);
-
-                // Show snackbar for item added
-                setSnackbar({ open: true, message: "Item added to wishlist", severity: "success" });
-            }
-        } catch (error) {
-            console.error("Error adding/removing item to/from wishlist:", error);
-        }
-    };
-
 
     const handleItemClick = (product) => {
         navigate(`/product/${product.id}`, { state: { product } });
     };
-
 
     return (
         <div className="new-arrivals">
@@ -322,8 +100,8 @@ export default function Arrivals() {
                 </div>
                 {visibleItems.map((item, index) => (
                     item && item.image_url ? (
-                        <div key={index} className="new-arrivals-product" >
-                <div className="product-image" onClick={() => handleItemClick(item) }>
+                        <div key={index} className="new-arrivals-product">
+                            <div className="product-image" onClick={() => handleItemClick(item)}>
                                 <img src={item.image_url} alt={`Product ${index}`} />
                             </div>
                             <div className="product-info">
@@ -332,14 +110,14 @@ export default function Arrivals() {
                                     <Typography className="current-price">{item.price}</Typography>
                                 </div>
                                 <div className="product-icons">
-                                    <IconButton onClick={() => handleFavoriteToggle(item)}>
+                                    <IconButton onClick={() => handleAddToWishlistToggle(item)}>
                                         {wishlist.some((wishlistItem) => wishlistItem.product_id === item.id) ? (
                                             <FavoriteOutlinedIcon />
                                         ) : (
                                             <FavoriteBorderOutlinedIcon />
                                         )}
                                     </IconButton>
-                                    <IconButton onClick={() => handleCartToggle(item)}>
+                                    <IconButton onClick={() => handleAddToCartToggle(item)}>
                                         {cart.some((cartItem) => cartItem.product_id === item.id) ? (
                                             <ShoppingCartOutlinedIcon />
                                         ) : (
@@ -360,11 +138,11 @@ export default function Arrivals() {
             </Button>
             <Snackbar
                 open={snackbar.open}
-                autoHideDuration={3000}
-                onClose={handleCloseSnackbar}
+                autoHideDuration={6000}
+                onClose={handleCloseSnackbar(setSnackbar)}
                 anchorOrigin={{ vertical: 'bottom', horizontal: 'left' }}
             >
-                <Alert onClose={handleCloseSnackbar} severity={snackbar.severity} sx={{ width: '100%' }}>
+                <Alert onClose={handleCloseSnackbar(setSnackbar)} severity={snackbar.severity}>
                     {snackbar.message}
                 </Alert>
             </Snackbar>
